Use MaaS token and API URL
The MaaS playbooks can either use a set of public cloud credentials or a
token and API URL. Customer deployments always require a token and API
URL. This commit adjusts how MaaS is deployed so that the existing
public cloud credentials are used to generate a token and API URL which
are then used by the playbooks instead.

This changes applies to all jobs deploying MaaS. Given that the use of
public cloud credentials in the playbooks is only for the benefit of
developers it makes sense for that to be the untested path.

The LEM AIO job has been updated to include the required MaaS stages.

This change does not directly address multi-physical-node deployments
using LEM however that can build on this work when it becomes a reality.

Connected https://github.com/rcbops/u-suk-dev/issues/1370